NMR-204 Fix calibration OOME and significantly improve performance
* removed unnecessary back and forth boxing/unboxing of data
* fixed start/stop position in getIntendedPeakPosition() and changed to
new findIndexOfValue() method
* added settings object to adjust quality control cycles (if necessary)
* updated to static access of methods from UtilityFunctions where
applicable

+++ openchrom/plugins/net.openchrom.nmr.processing.supplier.base/src/net/openchrom/nmr/processing/supplier/base/settings/support/IcoShiftAlignmentUtilities.java
@@ -110,7 +110,7 @@ public class IcoShiftAlignmentUtilities {
 			double[] dataArray = matrix.getData();
 			AcquisitionParameter parameter = measurement.getAcquisitionParameter();
 			for(int i = 0; i < dataArray.length; i++) {
-				newSignals.add(new IcoShiftSignal(parameter.toHz(BigDecimal.valueOf(chemicalShiftAxis[i].doubleValue())), dataArray[i]));
+				newSignals.add(new IcoShiftSignal(parameter.toHz(chemicalShiftAxis[i]), dataArray[i]));
 			}
 			alignmentResultIndex++;
 			filteredSpectrumMeasurement.setSignals(newSignals);
